refactor(client): organize public action types
