+++ src/client/HttpClient.ts
@@ -1,13 +1,17 @@
 import * as Client from './Client.js'
 import * as HttpTransport from './transports/HttpTransport.js'
-import type { ClientDefaults, Commands, HttpClient } from './types.js'
 
-export type { HttpClient }
+/** HTTP client instance. */
+export type HttpClient<
+  commands = Client.Commands,
+  defaults extends Client.Defaults = {},
+> = Client.Client<commands, HttpTransport.HttpTransport, defaults>
 
 /** Creates an HTTP typed client. */
-export function create<const commands = Commands, const defaults extends ClientDefaults = {}>(
-  options: HttpTransport.Options & defaults & ClientDefaults,
-): HttpClient<commands, defaults> {
+export function create<
+  const commands = Client.Commands,
+  const defaults extends Client.Defaults = {},
+>(options: HttpTransport.Options & defaults & Client.Defaults): HttpClient<commands, defaults> {
   const { baseUrl, fetch, headers, ...defaults } = options
   return Client.create<commands, HttpTransport.HttpTransport, defaults>({
     ...defaults,
