+++ src/client/index.ts
@@ -6,5 +6,6 @@ export * as MemoryClient from './MemoryClient.js'
 export * as MemoryTransport from './transports/MemoryTransport.js'
 export * as Resources from './Resources.js'
 export * as Rpc from './Rpc.js'
+export * as Run from './Run.js'
 export * as Transport from './transports/Transport.js'
-export type { Register } from './types.js'
+export type { Register } from './Client.js'
